app/agent/langraph_agent.py
from typing import Dict, List, Any
import os
from langgraph.graph import StateGraph, END
from langgraph.prebuilt import create_agent_executor
from langchain_openai import ChatOpenAI
from langchain.prompts import ChatPromptTemplate
from langchain.tools import Tool
from pydantic import BaseModel


class AgentState(BaseModel):
    query: str
    context: Dict[str, Any] = {}
    database_info: Dict[str, Any] = {}
    current_plan: List[str] = []
    execution_result: Dict[str, Any] = {}
    response: str = ""
    error: str = ""


def initialize_agent():
    """Initialize the LangGraph agent with tools and memory management."""

    # Initialize the language model
    llm = ChatOpenAI(
        model="gpt-4-turbo",
        temperature=0
    )

    # Define the system message for the agent
    system_message = """You are a database assistant that helps users interact with databases through natural language.
    You can interpret user queries, develop plans to retrieve or modify data, and execute those plans.

    You have access to the following database information:
    {database_info}

    Your goal is to understand what the user wants to do with the database and help them accomplish it.
    """

    # Define the agent's workflow as a graph
    workflow = StateGraph(AgentState)

    # Define the nodes

    # 1. Query understanding node
    def understand_query(state: AgentState) -> AgentState:
        """Parse and understand the user's natural language query."""
        prompt = ChatPromptTemplate.from_messages([
            ("system", system_message),
            ("user", "{query}")
        ])

        chain = prompt | llm

        result = chain.invoke({
            "query": state.query,
            "database_info": state.database_info
        })

        # Update state with query understanding
        state.context["understood_intent"] = result.content
        return state

    # 2. Query planning node
    def plan_execution(state: AgentState) -> AgentState:
        """Develop a plan for database operations based on the understood query."""
        # Implementation details
        return state

    # 3. Query execution node
    def execute_plan(state: AgentState) -> AgentState:
        """Execute the plan against the database."""
        # Implementation details
        return state

    # 4. Response formulation node
    def formulate_response(state: AgentState) -> AgentState:
        """Generate a natural language response based on execution results."""
        # Implementation details
        return state

    # 5. Error handling node
    def handle_error(state: AgentState) -> AgentState:
        """Handle any errors that occurred during processing."""
        # Implementation details
        return state

    # Add nodes to the graph
    workflow.add_node("understand_query", understand_query)
    workflow.add_node("plan_execution", plan_execution)
    workflow.add_node("execute_plan", execute_plan)
    workflow.add_node("formulate_response", formulate_response)
    workflow.add_node("handle_error", handle_error)

    # Define the edges of the graph
    workflow.add_edge("understand_query", "plan_execution")
    workflow.add_edge("plan_execution", "execute_plan")
    workflow.add_edge("execute_plan", "formulate_response")
    workflow.add_edge("formulate_response", END)

    # Add error handling edges
    workflow.add_conditional_edges(
        "understand_query",
        lambda state: "handle_error" if state.error else "plan_execution"
    )
    workflow.add_conditional_edges(
        "plan_execution",
        lambda state: "handle_error" if state.error else "execute_plan"
    )
    workflow.add_conditional_edges(
        "execute_plan",
        lambda state: "handle_error" if state.error else "formulate_response"
    )
    workflow.add_edge("handle_error", "formulate_response")

    # Set the entrypoint
    workflow.set_entry_point("understand_query")

    return workflow.compile()


def query_database(query: str, database_info: Dict[str, Any]) -> Dict[str, Any]:
    """Execute a query against the database using the LangGraph agent."""
    agent = initialize_agent()

    initial_state = AgentState(
        query=query,
        database_info=database_info
    )

    result = agent.invoke(initial_state)
    return {
        "response": result.response,
        "context": result.context,
        "execution_details": result.execution_result
    }
